+++ python/snewpy/models/ccsn.py
@@ -11,7 +11,6 @@ Updated summer 2020 by Jim Kneller & Arkin Worlikar. Subsequent updates
 provided by the SNEWS team.
 """
 
-from abc import abstractmethod, ABC
 from enum import IntEnum
 
 import astropy
@@ -41,251 +40,12 @@ try:
 except ImportError:
     pass
 
-from .neutrino import Flavor
-from .flavor_transformation import *
+from snewpy.neutrino import Flavor
+from snewpy.flavor_transformation import *
 
-
-def get_value(x):
-    """If quantity x has is an astropy Quantity with units, return just the
-    value.
-
-    Parameters
-    ----------
-    x : Quantity, float, or ndarray
-        Input quantity.
-
-    Returns
-    -------
-    value : float or ndarray
-    
-    :meta private:
-    """
-    if type(x) == Quantity:
-        return x.value
-    return x
-
-
-class SupernovaModel(ABC):
-    """Base class defining an interface to a supernova model."""
-    
-    def __init__(self):
-        pass
-
-    def __repr__(self):
-        if hasattr(self, "filename"):
-            # self.__class__ will be something like 
-            return f"{self.__class__.__name__}('{self.filename}')"
-        else:
-            return super().__repr__()
-
-    @abstractmethod
-    def get_time(self):
-        """Returns
-        -------
-            returns array of snapshot times from the simulation
-        """
-        pass
-
-    @abstractmethod
-    def get_initial_spectra(self, t, E, flavors=Flavor):
-        """Get neutrino spectra at the source.
-
-        Parameters
-        ----------
-        t : astropy.Quantity
-            Time to evaluate initial spectra.
-        E : astropy.Quantity or ndarray of astropy.Quantity
-            Energies to evaluate the initial spectra.
-        flavors: iterable of snewpy.neutrino.Flavor
-            Return spectra for these flavors only (default: all)
-
-        Returns
-        -------
-        initialspectra : dict
-            Dictionary of neutrino spectra, keyed by neutrino flavor.
-        """
-        pass
-
-    def get_initialspectra(self, *args):
-        """DO NOT USE! Only for backward compatibility!
-
-        :meta private:
-        """
-        print("[WARNING] Please use `get_initial_spectra()` instead of `get_initialspectra()`!")
-        return self.get_initial_spectra(*args)
-
-    def get_transformed_spectra(self, t, E, flavor_xform):
-        """Get neutrino spectra after applying oscillation.
-
-        Parameters
-        ----------
-        t : astropy.Quantity
-            Time to evaluate initial and oscillated spectra.
-        E : astropy.Quantity or ndarray of astropy.Quantity
-            Energies to evaluate the initial and oscillated spectra.
-        flavor_xform : FlavorTransformation
-            An instance from the flavor_transformation module.
-
-        Returns
-        -------
-        dict
-            Dictionary of transformed spectra, keyed by neutrino flavor.
-        """
-        initialspectra = self.get_initial_spectra(t, E)
-        transformed_spectra = {}
-
-        transformed_spectra[Flavor.NU_E] = \
-            flavor_xform.prob_ee(t, E) * initialspectra[Flavor.NU_E] + \
-            flavor_xform.prob_ex(t, E) * initialspectra[Flavor.NU_X]
-
-        transformed_spectra[Flavor.NU_X] = \
-            flavor_xform.prob_xe(t, E) * initialspectra[Flavor.NU_E] + \
-            flavor_xform.prob_xx(t, E) * initialspectra[Flavor.NU_X] 
-
-        transformed_spectra[Flavor.NU_E_BAR] = \
-            flavor_xform.prob_eebar(t, E) * initialspectra[Flavor.NU_E_BAR] + \
-            flavor_xform.prob_exbar(t, E) * initialspectra[Flavor.NU_X_BAR]
-
-        transformed_spectra[Flavor.NU_X_BAR] = \
-            flavor_xform.prob_xebar(t, E) * initialspectra[Flavor.NU_E_BAR] + \
-            flavor_xform.prob_xxbar(t, E) * initialspectra[Flavor.NU_X_BAR] 
-
-        return transformed_spectra   
-
-
-    def get_oscillatedspectra(self, *args):
-        """DO NOT USE! Only for backward compatibility!
-
-        :meta private:
-        """
-        print("[WARNING] Please use `get_transformed_spectra()` instead of `get_oscillatedspectra()`!")
-        return self.get_transformed_spectra(*args)
-
-
-class _GarchingArchiveModel(SupernovaModel):
-    """Subclass that reads models in the format used in the [Garching Supernova Archive](https://wwwmpa.mpa-garching.mpg.de/ccsnarchive/)."""
-    def __init__(self, filename, eos='LS220'):
-        """Initialize model.
-
-        Parameters
-        ----------
-        filename : str
-            Absolute or relative path to file prefix, we add nue/nuebar/nux.
-        eos : string
-            Equation of state used in simulation.
-        """
-        self.time = {}
-        self.luminosity = {}
-        self.meanE = {}
-        self.pinch = {}
-
-        # Store model metadata.
-        self.filename = os.path.basename(filename)
-        self.EOS = eos
-        self.progenitor_mass = float( (self.filename.split('s'))[1].split('c')[0] )  * u.Msun
-
-        # Read through the several ASCII files for the chosen simulation and
-        # merge the data into one giant table.
-        mergtab = None
-        for flavor in Flavor:
-            _flav = Flavor.NU_X if flavor == Flavor.NU_X_BAR else flavor
-            _sfx = _flav.name.replace('_', '').lower()
-            _filename = '{}_{}_{}'.format(filename, eos, _sfx)
-            _lname  = 'L_{}'.format(flavor.name)
-            _ename  = 'E_{}'.format(flavor.name)
-            _e2name = 'E2_{}'.format(flavor.name)
-            _aname  = 'ALPHA_{}'.format(flavor.name)
-
-            simtab = Table.read(_filename,
-                                names=['TIME', _lname, _ename, _e2name],
-                                format='ascii')
-            simtab['TIME'].unit = 's'
-            simtab[_lname].unit = '1e51 erg/s'
-            simtab[_aname] = (2*simtab[_ename]**2 - simtab[_e2name]) / (simtab[_e2name] - simtab[_ename]**2)
-            simtab[_ename].unit = 'MeV'
-            del simtab[_e2name]
-
-            if mergtab is None:
-                mergtab = simtab
-            else:
-                mergtab = join(mergtab, simtab, keys='TIME', join_type='left')
-                mergtab[_lname].fill_value = 0.
-                mergtab[_ename].fill_value = 0.
-                mergtab[_aname].fill_value = 0.
-        simtab = mergtab.filled()
-
-        self.time = simtab['TIME'].to('s')
-
-        for flavor in Flavor:
-            # Set the dictionary of luminosity, mean energy, and shape
-            # parameter keyed by NU_E, NU_X, NU_E_BAR, NU_X_BAR.
-            _lname  = 'L_{}'.format(flavor.name)
-            self.luminosity[flavor] = simtab[_lname].to('erg/s')
-
-            _ename  = 'E_{}'.format(flavor.name)
-            self.meanE[flavor] = simtab[_ename].to('MeV')
-
-            _aname  = 'ALPHA_{}'.format(flavor.name)
-            self.pinch[flavor] = simtab[_aname]
-
-    def get_time(self):
-        """Get grid of model times.
-
-        Returns
-        -------
-        time : ndarray
-            Grid of times used in the model.
-        """
-        return self.time
-
-    def get_initial_spectra(self, t, E, flavors=Flavor):
-        """Get neutrino spectra/luminosity curves before oscillation.
-
-        Parameters
-        ----------
-        t : astropy.Quantity
-            Time to evaluate initial spectra.
-        E : astropy.Quantity or ndarray of astropy.Quantity
-            Energies to evaluate the initial spectra.
-        flavors: iterable of snewpy.neutrino.Flavor
-            Return spectra for these flavors only (default: all)
-
-        Returns
-        -------
-        initialspectra : dict
-            Dictionary of model spectra, keyed by neutrino flavor.
-        """
-        initialspectra = {}
-
-        # Avoid division by zero in energy PDF below.
-        E[E==0] = np.finfo(float).eps * E.unit
-
-        # Estimate L(t), <E_nu(t)> and alpha(t). Express all energies in erg.
-        E = E.to_value('erg')
-
-        # Make sure input time uses the same units as the model time grid, or
-        # the interpolation will not work correctly.
-        t = t.to(self.time.unit)
-
-        for flavor in flavors:
-            # Use np.interp rather than scipy.interpolate.interp1d because it
-            # can handle dimensional units (astropy.Quantity).
-            L  = get_value(np.interp(t, self.time, self.luminosity[flavor].to('erg/s')))
-            Ea = get_value(np.interp(t, self.time, self.meanE[flavor].to('erg')))
-            a  = np.interp(t, self.time, self.pinch[flavor])
-
-            if L==0:
-                initialspectra[flavor] = 0.0*E/(u.erg*u.s)
-            else:
-                # For numerical stability, evaluate log PDF and then exponentiate.
-                initialspectra[flavor] = \
-                  np.exp(np.log(L) - (2+a)*np.log(Ea) + (1+a)*np.log(1+a)
-                        - loggamma(1+a) + a*np.log(E) - (1+a)*(E/Ea)) / (u.erg * u.s)
-
-        return initialspectra
-
+from .base import SupernovaModel, _GarchingArchiveModel,PinchedModel
  
-class Analytic3Species(SupernovaModel):
+class Analytic3Species(PinchedModel):
     """Allow to generate an analytic model given total luminosity,
     average energy, and rms or pinch, for each species.
     """
@@ -317,59 +77,6 @@ class Analytic3Species(SupernovaModel):
             self.meanE[flavor] = simtab['E_{}'.format(_flav.name)] * u.MeV
             self.pinch[flavor] = simtab['ALPHA_{}'.format(_flav.name)]
 
-    def get_time(self):
-        """Get grid of model times.
-
-        Returns
-        -------
-        time : ndarray
-            Grid of times used in the model.
-        """
-        return self.time
-    
-    def get_initial_spectra(self, t, E, flavors=Flavor):
-        """Get neutrino spectra/luminosity curves after oscillation.
-
-        Parameters
-        ----------
-        t : astropy.Quantity
-            Time to evaluate initial spectra.
-        E : astropy.Quantity or ndarray of astropy.Quantity
-            Energies to evaluate the initial spectra.
-        flavors: iterable of snewpy.neutrino.Flavor
-            Return spectra for these flavors only (default: all)
-
-        Returns
-        -------
-        initialspectra : dict
-            Dictionary of model spectra, keyed by neutrino flavor.
-        """
-        initialspectra={}
-
-        # Avoid division by zero in energy PDF below.
-        E[E==0] = np.finfo(float).eps * E.unit
-
-        # Estimate L(t), <E_nu(t)> and alpha(t). Express all energies in erg.
-        E = E.to_value('erg')
-
-        # Make sure input time uses the same units as the model time grid, or
-        # the interpolation will not work correctly.
-        t = t.to(self.time.unit)
-
-        for flavor in flavors:
-            # Use np.interp rather than scipy.interpolate.interp1d because it
-            # can handle dimensional units (astropy.Quantity).
-            L  = get_value(np.interp(t, self.time, self.luminosity[flavor].to('erg/s')))
-            Ea = get_value(np.interp(t, self.time, self.meanE[flavor].to('erg')))
-            a  = np.interp(t, self.time, self.pinch[flavor])
-
-            # For numerical stability, evaluate log PDF and then exponentiate.
-            initialspectra[flavor] = \
-                np.exp(np.log(L) - (2+a)*np.log(Ea) + (1+a)*np.log(1+a)
-                       - loggamma(1+a) + a*np.log(E) - (1+a)*(E/Ea)) / (u.erg * u.s)
-
-        return initialspectra
-
     def __str__(self):
         """Default representation of the model.
         """
@@ -383,7 +90,7 @@ class Analytic3Species(SupernovaModel):
         return mod  
 
 
-class Nakazato_2013(SupernovaModel):
+class Nakazato_2013(PinchedModel):
     """Set up a model based on simulations from Nakazato et al., ApJ S 205:2
     (2013), ApJ 804:75 (2015), PASJ 73:639 (2021). See also http://asphwww.ph.noda.tus.ac.jp/snn/.
     """
@@ -431,59 +138,6 @@ class Nakazato_2013(SupernovaModel):
             self.meanE[flavor] = simtab['E_{}'.format(_flav.name)].to('MeV')
             self.pinch[flavor] = simtab['ALPHA_{}'.format(_flav.name)]
 
-    def get_time(self):
-        """Get grid of model times.
-
-        Returns
-        -------
-        time : ndarray
-            Grid of times used in the model.
-        """
-        return self.time
-    
-    def get_initial_spectra(self, t, E, flavors=Flavor):
-        """Get neutrino spectra/luminosity at the source.
-
-        Parameters
-        ----------
-        t : astropy.Quantity
-            Time to evaluate initial spectra.
-        E : astropy.Quantity or ndarray of astropy.Quantity
-            Energies to evaluate the initial spectra.
-        flavors: iterable of snewpy.neutrino.Flavor
-            Return spectra for these flavors only (default: all)
-
-        Returns
-        -------
-        initialspectra : dict
-            Dictionary of model spectra, keyed by neutrino flavor.
-        """
-        initialspectra = {}
-
-        # Avoid division by zero in energy PDF below.
-        E[E==0] = np.finfo(float).eps * E.unit
-
-        # Estimate L(t), <E_nu(t)> and alpha(t). Express all energies in erg.
-        E = E.to_value('erg')
-
-        # Make sure input time uses the same units as the model time grid, or
-        # the interpolation will not work correctly.
-        t = t.to(self.time.unit)
-
-        for flavor in flavors:
-            # Use np.interp rather than scipy.interpolate.interp1d because it
-            # can handle dimensional units (astropy.Quantity).
-            L  = get_value(np.interp(t, self.time, self.luminosity[flavor].to('erg/s')))
-            Ea = get_value(np.interp(t, self.time, self.meanE[flavor].to('erg')))
-            a  = np.interp(t, self.time, self.pinch[flavor])
-
-            # For numerical stability, evaluate log PDF and then exponentiate.
-            initialspectra[flavor] = \
-                np.exp(np.log(L) - (2+a)*np.log(Ea) + (1+a)*np.log(1+a) 
-                       - loggamma(1+a) + a*np.log(E) - (1+a)*(E/Ea)) / (u.erg * u.s)
-
-        return initialspectra
-
     def __str__(self):
         """Default representation of the model.
         """
@@ -509,7 +163,7 @@ class Nakazato_2013(SupernovaModel):
         return mod + '\n'.join(s)
 
 
-class Sukhbold_2015(SupernovaModel):
+class Sukhbold_2015(PinchedModel):
     """Set up a model based on simulations from Sukhbold et al., ApJ 821:38,2016. Models were shared privately by email.
     """
 
@@ -543,57 +197,6 @@ class Sukhbold_2015(SupernovaModel):
             self.meanE[flavor] = simtab['E_{}'.format(flavor.name)].to('MeV')
             self.pinch[flavor] = simtab['ALPHA_{}'.format(flavor.name)]
             
-    def get_time(self):
-        """Get grid of model times.
-
-        Returns
-        -------
-        time : ndarray
-            Grid of times used in the model.
-        """
-        return self.time
-    
-    def get_initial_spectra(self, t, E, flavors=Flavor):
-        """Get neutrino spectra/luminosity curves after oscillation.
-
-        Parameters
-        ----------
-        t : astropy.Quantity
-            Time to evaluate initial spectra.
-        E : astropy.Quantity or ndarray of astropy.Quantity
-            Energies to evaluate the initial spectra.
-        flavors: iterable of snewpy.neutrino.Flavor
-            Return spectra for these flavors only (default: all)
-
-        Returns
-        -------
-        initialspectra : dict
-            Dictionary of model spectra, keyed by neutrino flavor.
-        """
-        initialspectra = {}
-
-        # Avoid division by zero in energy PDF below.
-        E[E==0] = np.finfo(float).eps * E.unit
-
-        # Estimate L(t), <E_nu(t)>, and alpha(t). Express all energies in erg.
-        E = E.to_value('erg')
-
-        # Make sure input time uses the same units as the global time grid or
-        # the interpolation will not work properly.
-        t = t.to(self.time.unit)
-
-        for flavor in flavors:
-            L  = get_value(np.interp(t, self.time, self.luminosity[flavor].to('erg/s')))
-            Ea = get_value(np.interp(t, self.time, self.meanE[flavor].to('erg')))
-            a  = np.interp(t, self.time, self.pinch[flavor])
-
-            # For numerical stability, evaluate log PDF then exponentiate.
-            initialspectra[flavor] = \
-                np.exp(np.log(L) - (2+a)*np.log(Ea) + (1+a)*np.log(1+a) 
-                       - loggamma(1+a) + a*np.log(E) - (1+a)*(E/Ea)) / (u.erg * u.s)
-
-        return initialspectra
-
     def __str__(self):
         """Default representation of the model.
         """
@@ -717,7 +320,7 @@ class Walk_2019(_GarchingArchiveModel):
              ]
         return mod + '\n'.join(s)
 
-class OConnor_2013(SupernovaModel):
+class OConnor_2013(PinchedModel):
     """Set up a model based on the black hole formation simulation in O'Connor & Ott (2013). 
     """
     def __init__(self, base, mass=15, eos='LS220'):
@@ -771,59 +374,6 @@ class OConnor_2013(SupernovaModel):
             self.meanE[flavor] = simtab['E_{}'.format(_flav.name)] * u.MeV
             self.pinch[flavor] = simtab['ALPHA_{}'.format(_flav.name)]
 
-    def get_time(self):
-        """Get grid of model times.
-
-        Returns
-        -------
-        time : ndarray
-            Grid of times used in the model.
-        """
-        return self.time
-
-    def get_initial_spectra(self, t, E, flavors=Flavor):
-        """Get neutrino spectra/luminosity curves before oscillation.
-
-        Parameters
-        ----------
-        t : astropy.Quantity
-            Time to evaluate initial spectra.
-        E : astropy.Quantity or ndarray of astropy.Quantity
-            Energies to evaluate the initial spectra.
-        flavors: iterable of snewpy.neutrino.Flavor
-            Return spectra for these flavors only (default: all)
-
-        Returns
-        -------
-        initialspectra : dict
-            Dictionary of model spectra, keyed by neutrino flavor.
-        """
-        initialspectra = {}
-
-        # Avoid division by zero in energy PDF below.
-        E[E==0] = np.finfo(float).eps * E.unit
-
-        # Estimate L(t), <E_nu(t)> and alpha(t). Express all energies in erg.
-        E = E.to_value('erg')
-
-        # Make sure input time uses the same units as the model time grid, or
-        # the interpolation will not work correctly.
-        t = t.to(self.time.unit)
-
-        for flavor in flavors:
-            # Use np.interp rather than scipy.interpolate.interp1d because it
-            # can handle dimensional units (astropy.Quantity).
-            L  = get_value(np.interp(t, self.time, self.luminosity[flavor].to('erg/s')))
-            Ea = get_value(np.interp(t, self.time, self.meanE[flavor].to('erg')))
-            a  = np.interp(t, self.time, self.pinch[flavor])
-
-            # For numerical stability, evaluate log PDF and then exponentiate.
-            initialspectra[flavor] = \
-                np.exp(np.log(L) - (2+a)*np.log(Ea) + (1+a)*np.log(1+a)
-                       - loggamma(1+a) + a*np.log(E) - (1+a)*(E/Ea)) / (u.erg * u.s)
-
-        return initialspectra
-
     def __str__(self):
         """Default representation of the model.
         """
@@ -844,7 +394,7 @@ class OConnor_2013(SupernovaModel):
              ]
         return mod + '\n'.join(s)
 
-class OConnor_2015(SupernovaModel):
+class OConnor_2015(PinchedModel):
     """Set up a model based on the black hole formation simulation in O'Connor (2015). 
     """
     def __init__(self, filename, eos='LS220'):
@@ -896,59 +446,6 @@ class OConnor_2015(SupernovaModel):
             self.meanE[flavor] = simtab['E_{}'.format(_flav.name)] * u.MeV
             self.pinch[flavor] = simtab['ALPHA_{}'.format(_flav.name)]
 
-    def get_time(self):
-        """Get grid of model times.
-
-        Returns
-        -------
-        time : ndarray
-            Grid of times used in the model.
-        """
-        return self.time
-
-    def get_initial_spectra(self, t, E, flavors=Flavor):
-        """Get neutrino spectra/luminosity curves before oscillation.
-
-        Parameters
-        ----------
-        t : astropy.Quantity
-            Time to evaluate initial spectra.
-        E : astropy.Quantity or ndarray of astropy.Quantity
-            Energies to evaluate the initial spectra.
-        flavors: iterable of snewpy.neutrino.Flavor
-            Return spectra for these flavors only (default: all)
-
-        Returns
-        -------
-        initialspectra : dict
-            Dictionary of model spectra, keyed by neutrino flavor.
-        """
-        initialspectra = {}
-
-        # Avoid division by zero in energy PDF below.
-        E[E==0] = np.finfo(float).eps * E.unit
-
-        # Estimate L(t), <E_nu(t)> and alpha(t). Express all energies in erg.
-        E = E.to_value('erg')
-
-        # Make sure input time uses the same units as the model time grid, or
-        # the interpolation will not work correctly.
-        t = t.to(self.time.unit)
-
-        for flavor in flavors:
-            # Use np.interp rather than scipy.interpolate.interp1d because it
-            # can handle dimensional units (astropy.Quantity).
-            L  = get_value(np.interp(t, self.time, self.luminosity[flavor].to('erg/s')))
-            Ea = get_value(np.interp(t, self.time, self.meanE[flavor].to('erg')))
-            a  = np.interp(t, self.time, self.pinch[flavor])
-
-            # For numerical stability, evaluate log PDF and then exponentiate.
-            initialspectra[flavor] = \
-                np.exp(np.log(L) - (2+a)*np.log(Ea) + (1+a)*np.log(1+a)
-                       - loggamma(1+a) + a*np.log(E) - (1+a)*(E/Ea)) / (u.erg * u.s)
-
-        return initialspectra
-
     def __str__(self):
         """Default representation of the model.
         """
@@ -969,7 +466,7 @@ class OConnor_2015(SupernovaModel):
              ]
         return mod + '\n'.join(s)
 
-class Zha_2021(SupernovaModel):
+class Zha_2021(PinchedModel):
     """Set up a model based on the hadron-quark phse transition models from Zha et al. 2021. 
     """
     def __init__(self, filename, eos='STOS_B145'):
@@ -1029,59 +526,6 @@ class Zha_2021(SupernovaModel):
             self.meanE[flavor] = simtab['E_{}'.format(_flav.name)] * u.MeV
             self.pinch[flavor] = simtab['ALPHA_{}'.format(_flav.name)]
 
-    def get_time(self):
-        """Get grid of model times.
-
-        Returns
-        -------
-        time : ndarray
-            Grid of times used in the model.
-        """
-        return self.time
-
-    def get_initial_spectra(self, t, E, flavors=Flavor):
-        """Get neutrino spectra/luminosity curves before oscillation.
-
-        Parameters
-        ----------
-        t : astropy.Quantity
-            Time to evaluate initial spectra.
-        E : astropy.Quantity or ndarray of astropy.Quantity
-            Energies to evaluate the initial spectra.
-        flavors: iterable of snewpy.neutrino.Flavor
-            Return spectra for these flavors only (default: all)
-
-        Returns
-        -------
-        initialspectra : dict
-            Dictionary of model spectra, keyed by neutrino flavor.
-        """
-        initialspectra = {}
-
-        # Avoid division by zero in energy PDF below.
-        E[E==0] = np.finfo(float).eps * E.unit
-
-        # Estimate L(t), <E_nu(t)> and alpha(t). Express all energies in erg.
-        E = E.to_value('erg')
-
-        # Make sure input time uses the same units as the model time grid, or
-        # the interpolation will not work correctly.
-        t = t.to(self.time.unit)
-
-        for flavor in flavors:
-            # Use np.interp rather than scipy.interpolate.interp1d because it
-            # can handle dimensional units (astropy.Quantity).
-            L  = get_value(np.interp(t, self.time, self.luminosity[flavor].to('erg/s')))
-            Ea = get_value(np.interp(t, self.time, self.meanE[flavor].to('erg')))
-            a  = np.interp(t, self.time, self.pinch[flavor])
-
-            # For numerical stability, evaluate log PDF and then exponentiate.
-            initialspectra[flavor] = \
-                np.exp(np.log(L) - (2+a)*np.log(Ea) + (1+a)*np.log(1+a)
-                       - loggamma(1+a) + a*np.log(E) - (1+a)*(E/Ea)) / (u.erg * u.s)
-
-        return initialspectra
-
     def __str__(self):
         """Default representation of the model.
         """
@@ -1103,7 +547,7 @@ class Zha_2021(SupernovaModel):
         return mod + '\n'.join(s)    
 
 
-class Warren_2020(SupernovaModel):
+class Warren_2020(PinchedModel):
     """Set up a model based on simulations from Warren et al., ApJ 898:139, 2020.
     Neutrino fluxes available at https://doi.org/10.5281/zenodo.3667908."""
 
@@ -1165,64 +609,6 @@ class Warren_2020(SupernovaModel):
             self.meanE[flavor] = simtab['E_{}'.format(_flav.name)] * u.MeV
             self.pinch[flavor] = simtab['ALPHA_{}'.format(_flav.name)]
 
-    def get_time(self):
-        """Get grid of model times.
-
-        Returns
-        -------
-        time : ndarray
-            Grid of times used in the model.
-        """
-        return self.time
-
-    def get_initial_spectra(self, t, E, flavors=Flavor):
-        """Get neutrino spectra/luminosity curves before oscillation.
-
-        Parameters
-        ----------
-        t : astropy.Quantity
-            Time to evaluate initial spectra.
-        E : astropy.Quantity or ndarray of astropy.Quantity
-            Energies to evaluate the initial spectra.
-        flavors: iterable of snewpy.neutrino.Flavor
-            Return spectra for these flavors only (default: all)
-
-        Returns
-        -------
-        initialspectra : dict
-            Dictionary of model spectra, keyed by neutrino flavor.
-        """
-        initialspectra = {}
-
-        # Avoid division by zero in energy PDF below.
-        E[E==0] = np.finfo(float).eps * E.unit
-
-        # Estimate L(t), <E_nu(t)> and alpha(t). Express all energies in erg.
-        E = E.to_value('erg')
-
-        # Make sure input time uses the same units as the model time grid, or
-        # the interpolation will not work correctly.
-        t = t.to(self.time.unit)
-
-        for flavor in flavors:
-            # Use np.interp rather than scipy.interpolate.interp1d because it
-            # can handle dimensional units (astropy.Quantity).
-            L  = get_value(np.interp(t, self.time, self.luminosity[flavor].to('erg/s')))
-            Ea = get_value(np.interp(t, self.time, self.meanE[flavor].to('erg')))
-            a  = np.interp(t, self.time, self.pinch[flavor])
-
-            # Sanity check to avoid invalid values of Ea, alpha, and L.
-            initialspectra[flavor] = np.zeros_like(E, dtype=float) / (u.erg*u.s)
-            if L <= 0. or Ea <= 0. or a <= -2.:
-                continue
-
-            # For numerical stability, evaluate log PDF and then exponentiate.
-            initialspectra[flavor] = \
-                np.exp(np.log(L) - (2+a)*np.log(Ea) + (1+a)*np.log(1+a)
-                       - loggamma(1+a) + a*np.log(E) - (1+a)*(E/Ea)) / (u.erg * u.s)
-
-        return initialspectra
-
     def __str__(self):
         """Default representation of the model.
         """
@@ -1246,7 +632,7 @@ class Warren_2020(SupernovaModel):
         return mod + '\n'.join(s)
 
 
-class Kuroda_2020(SupernovaModel):
+class Kuroda_2020(PinchedModel):
     """Set up a model based on simulations from Kuroda et al. (2020)."""
 
     def __init__(self, filename, eos='LS220', mass=20*u.Msun):
@@ -1290,59 +676,6 @@ class Kuroda_2020(SupernovaModel):
 
             # There is no pinch parameter so use alpha=2.0.
             self.pinch[flavor] = np.full_like(self.meanE[flavor].value, 2.)
-
-    def get_time(self):
-        """Get grid of model times.
-
-        Returns
-        -------
-        time : ndarray
-            Grid of times used in the model.
-        """
-        return self.time
-
-    def get_initial_spectra(self, t, E, flavors=Flavor):
-        """Get neutrino spectra/luminosity curves before oscillation.
-
-        Parameters
-        ----------
-        t : astropy.Quantity
-            Time to evaluate initial spectra.
-        E : astropy.Quantity or ndarray of astropy.Quantity
-            Energies to evaluate the initial spectra.
-        flavors: iterable of snewpy.neutrino.Flavor
-            Return spectra for these flavors only (default: all)
-
-        Returns
-        -------
-        initialspectra : dict
-            Dictionary of model spectra, keyed by neutrino flavor.
-        """
-        initialspectra = {}
-
-        # Avoid division by zero in energy PDF below.
-        E[E==0] = np.finfo(float).eps * E.unit
-
-        # Estimate L(t), <E_nu(t)> and alpha(t). Express all energies in erg.
-        E = E.to_value('erg')
-
-        # Make sure input time uses the same units as the model time grid, or
-        # the interpolation will not work correctly.
-        t = t.to(self.time.unit)
-
-        for flavor in flavors:
-            # Use np.interp rather than scipy.interpolate.interp1d because it
-            # can handle dimensional units (astropy.Quantity).
-            L  = get_value(np.interp(t, self.time, self.luminosity[flavor].to('erg/s')))
-            Ea = get_value(np.interp(t, self.time, self.meanE[flavor].to('erg')))
-            a  = np.interp(t, self.time, self.pinch[flavor])
-
-            # For numerical stability, evaluate log PDF and then exponentiate.
-            initialspectra[flavor] = \
-                np.exp(np.log(L) - (2+a)*np.log(Ea) + (1+a)*np.log(1+a)
-                       - loggamma(1+a) + a*np.log(E) - (1+a)*(E/Ea)) / (u.erg * u.s)
-
-        return initialspectra
 
     def __str__(self):
         """Default representation of the model.
